services/llm/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
transformers==4.38.2
torch==2.2.1
sentencepiece==0.2.0
typing-extensions>=4.7.0
opea-comps==1.2 